% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgt-package.R
\docType{package}
\name{dgt-package}
\alias{dgt}
\alias{dgt-package}
\title{dgt: Distributional Generalizability Theory}
\description{
Extends classical Generalizability Theory (G-theory) to non-Gaussian
measurements. Computes response-scale ICCs, information-theoretic
reliability, hurdle model decompositions, D-study curves, and
overestimation diagnostics from brms model posteriors.
}
\section{Main Functions}{

\describe{
\item{\code{\link{dgt_icc}}}{Compute response-scale, link-scale,
and information ICCs.}
\item{\code{\link{dgt_dstudy}}}{Compute D-study curves with
credible bands.}
\item{\code{\link{dgt_variance}}}{Five-component variance
decomposition for hurdle models.}
\item{\code{\link{dgt_required_n}}}{Required occasions for a
target reliability.}
\item{\code{\link{dgt_overestimation}}}{Overestimation and D-study
ratios (Theorem 6).}
}
}

\section{Supported Families}{

\itemize{
\item \code{gaussian} (with log-transformed response)
\item \code{lognormal}
\item \code{hurdle_lognormal}
\item \code{hurdle_poisson}, \code{hurdle_negbinomial} (hurdle counts
with a reference exposure; see \code{\link{dgt_icc}} and
\code{\link{dgt_hurdle_count_population}})
\item \code{poisson}, \code{negbinomial}, \code{bernoulli}, \code{binomial}
}
}

\section{Key Theorems}{

\itemize{
\item \strong{Theorem 1}: Lognormal response-scale ICC =
(exp(s2p)-1)/(exp(s2eta)-1)
\item \strong{Theorem 1a}: ICC_Y < ICC_eta always
(classical G-theory overestimates)
\item \strong{Theorem 2}: D-study = Spearman-Brown(ICC_Y)
\item \strong{Theorem 3}: Hurdle composite ICC with five-component
variance decomposition
\item \strong{Theorem 5}: Information ICC = link-scale ICC for
invertible links
\item \strong{Theorem 6}: ICC_Y < ICC_eta = ICC_I (three-ICC ordering)
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rckaruna/dgt}
  \item Report bugs at \url{https://github.com/rckaruna/dgt/issues}
}

}
\author{
\strong{Maintainer}: Ruwan Karunanayaka \email{ruwan.karunanayaka@ufv.ca} (\href{https://orcid.org/0009-0008-2459-1569}{ORCID})

Authors:
\itemize{
  \item Ruwan Karunanayaka \email{ruwan.karunanayaka@ufv.ca} (\href{https://orcid.org/0009-0008-2459-1569}{ORCID})
}

}
