% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link-utils.R
\name{.binary_link}
\alias{.binary_link}
\title{Link function of a bernoulli / binomial brms fit}
\usage{
.binary_link(fit)
}
\arguments{
\item{fit}{A brms fit.}
}
\value{
Character scalar: "logit", "probit", "cloglog", or "cauchit".
}
\description{
Link function of a bernoulli / binomial brms fit
}
\keyword{internal}
