% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.detect_family}
\alias{.detect_family}
\title{Detect the distributional family from a brms fit}
\usage{
.detect_family(fit)
}
\arguments{
\item{fit}{A brms model fit.}
}
\value{
Character string: "lognormal", "hurdle_lognormal", "gaussian",
or the raw family name.
}
\description{
Detect the distributional family from a brms fit
}
\keyword{internal}
