% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstudy.R
\name{.find_n_threshold}
\alias{.find_n_threshold}
\title{Find first n where median D-study >= threshold}
\usage{
.find_n_threshold(mat, n_grid, threshold)
}
\description{
Find first n where median D-study >= threshold
}
\keyword{internal}
