% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link-utils.R
\name{.inv_link_fun}
\alias{.inv_link_fun}
\title{Inverse link function for a given link name}
\usage{
.inv_link_fun(link)
}
\arguments{
\item{link}{Character. One of "logit", "probit", "cloglog", "cauchit".}
}
\value{
A vectorised function mapping the linear predictor to a
probability.
}
\description{
Inverse link function for a given link name
}
\keyword{internal}
