% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc-hurdle-count.R
\name{.q0_hurdle_count}
\alias{.q0_hurdle_count}
\title{Zero probability of the untruncated count family}
\usage{
.q0_hurdle_count(mu, family, phi = Inf)
}
\description{
Zero probability of the untruncated count family
}
\keyword{internal}
