% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstudy.R
\name{.required_n_bernoulli}
\alias{.required_n_bernoulli}
\title{Compute required n for Bernoulli D-study draw matrices}
\usage{
.required_n_bernoulli(ds_draws, n_grid, thresholds)
}
\description{
Compute required n for Bernoulli D-study draw matrices
}
\keyword{internal}
