% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstudy.R
\name{.required_n_from_draws}
\alias{.required_n_from_draws}
\title{Compute required n from D-study draw matrices}
\usage{
.required_n_from_draws(ds_draws, n_grid, thresholds)
}
\description{
Compute required n from D-study draw matrices
}
\keyword{internal}
