% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstudy.R
\name{.summarize_dstudy_matrix}
\alias{.summarize_dstudy_matrix}
\title{Summarize a D-study draw matrix into median + CI}
\usage{
.summarize_dstudy_matrix(mat, n_grid, type, probs)
}
\description{
Summarize a D-study draw matrix into median + CI
}
\keyword{internal}
