% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.R
\name{plot.dgt_allocate}
\alias{plot.dgt_allocate}
\title{Plot a dgt_allocate standard-error surface}
\usage{
\method{plot}{dgt_allocate}(x, scale = c("response", "link"), target = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{"dgt_allocate"}.}

\item{scale}{Character. Which surface to draw, \code{"response"}
(default) or \code{"link"}.}

\item{target}{Numeric or NULL. Draw a contour at this standard error.}

\item{...}{Unused.}
}
\description{
Plot a dgt_allocate standard-error surface
}
