% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.dgt_icc}
\alias{plot.dgt_icc}
\title{Plot ICC posterior distributions}
\usage{
\method{plot}{dgt_icc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"dgt_icc"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A ggplot2 object.
}
\description{
Plot ICC posterior distributions
}
