% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.R
\name{print.dgt_allocate}
\alias{print.dgt_allocate}
\title{Print a dgt_allocate object}
\usage{
\method{print}{dgt_allocate}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"dgt_allocate"}.}

\item{...}{Unused.}
}
\description{
Print a dgt_allocate object
}
